pyproject.toml
[project]
name = "modulo-agendamento"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "langchain-core>=0.3.63",
    "langchain-openai==0.3.16",
    "langgraph>=0.4.7",
    "langgraph-checkpoint-mongodb>=0.1.3",
    "langgraph-cli[inmem]>=0.2.10",
    "langsmith>=0.3.43",
    "pydantic-settings>=2.9.1",
    "pymongo>=4.11.3",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
